Move LSP server verification test to ExpertTest only
ElixirLS requires a proper Elixir project environment (with mix.exs) to
initialize properly, which is not available in the test environment.
Expert is a standalone binary that doesn't have this requirement.

- Remove verification test from BaseLspServiceTestCase
- Add verification test to ExpertTest only
- Add explanatory comment to ElixirLSTest
