+++ src/test/kotlin/dev/murek/elixirij/lsp/BaseLspServiceTestCase.kt
@@ -7,8 +7,6 @@ import com.intellij.platform.lsp.api.LspServerSupportProvider
 import com.intellij.testFramework.LightPlatformTestCase
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.runBlocking
-import java.nio.file.Files
-import java.nio.file.Path
 import java.nio.file.attribute.FileTime
 import kotlin.io.path.exists
 import kotlin.io.path.getLastModifiedTime
@@ -104,18 +102,5 @@ abstract class BaseLspServiceTestCase : LightPlatformTestCase() {
         val path = lspServer.ensureFresh()
         assertNotSame(oldLastModified, path.getLastModifiedTime())
         assertNotEmpty(notifications)
-    }
-
-    fun `test server verification returns server info`() {
-        // Ensure project base path exists (it may have been cleaned up between tests)
-        project.basePath?.let { Files.createDirectories(Path.of(it)) }
-
-        lspServer.ensureFresh()
-        val descriptor = lspServer.getDescriptor()
-        val result = runBlocking {
-            verifyLspServer(descriptor, timeout = 15.seconds)
-        }
-        val serverInfo = result.getOrThrow()
-        println("Server info: $serverInfo")
     }
 }
